Added extended graphics support

+++ include/chip8.h
@@ -18,7 +18,7 @@ typedef enum {
 int load_program(const char* program_path);
 void print_state();
 void init_chip8();
-unsigned short next_cycle();
+unsigned int next_cycle();
 unsigned char* get_video_mem();
 Flag decrement_timers();
 void skip_key(unsigned char reg, bool is_equal, unsigned char key);
